+++ src/pages/Background/index.js
@@ -1,5 +1,6 @@
 import '../../assets/img/icon-34.png';
 import '../../assets/img/icon-128.png';
+import '../../assets/img/icon-128-eye.png';
 
 let sidebarOpen = true; // open -> true  |  close -> false
 let sidebarScrollPosition = {
@@ -16,15 +17,23 @@ chrome.storage.local.get(['sidebarOpen'], (result) => {
 
 const changeBrowserIconBadgeWithSidebarOpenStatus = (status) => {
   if (status) {
-    chrome.browserAction.setBadgeText({
-      text: '👀',
-    });
-    chrome.browserAction.setBadgeBackgroundColor({
-      color: [255, 255, 255, 100],
+    // chrome.browserAction.setBadgeText({
+    //   text: '👀',
+    // });
+    // chrome.browserAction.setBadgeBackgroundColor({
+    //   color: [255, 255, 255, 100],
+    // });
+
+    chrome.browserAction.setIcon({
+      path: chrome.extension.getURL('icon-128-eye.png'),
     });
   } else {
-    chrome.browserAction.setBadgeText({
-      text: '',
+    // chrome.browserAction.setBadgeText({
+    //   text: '',
+    // });
+
+    chrome.browserAction.setIcon({
+      path: chrome.extension.getURL('icon-128.png'),
     });
   }
 };
@@ -118,6 +127,12 @@ chrome.browserAction.onClicked.addListener((senderTab) => {
   //   from: 'background',
   //   msg: 'TOGGLE_SIDEBAR',
   // });
+});
+
+chrome.commands.onCommand.addListener(function(command) {
+  if (command === 'toggle_sidebar_with_shortcut_1') {
+    toggleSidebar();
+  }
 });
 
 chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {